requirements.txt
accelerate==1.0.1
aesara==2.9.4
aiohttp==3.10.10
apex==0.9.10dev
astunparse==1.6.3
attr==0.3.2
audiocraft==1.3.0
av==13.1.0
azureml==0.2.7
beartype==0.19.0
beautifulsoup4==4.12.3
biotite==1.0.1
brotli==1.1.0
brotlicffi==1.1.0.0
causal_conv1d==1.4.0
classy_vision==0.7.0
clearml==1.16.5
codecarbon==2.7.2
comet_ml==3.47.1
conda==4.3.16
ConfigParser==7.1.0
contextlib2==21.6.0
coremltools==8.0
cryptography==43.0.3
ctypes_snappy==1.03
Cython==3.0.11
dagshub==0.3.41
dall_e==0.1
decord==0.6.0
deeplearning==0.0.1
diffusers==0.31.0
dill==0.3.9
dill==0.3.8
disco==1.40.4
dl==0.1.0
docutils==0.21.2
dvclive==3.48.0
essentia==2.1b6.dev1177
etcd==2.0.8
expecttest==0.2.1
fairseq==0.12.2
fairseq2==0.2.1
faiss==1.5.3
fastai==2.7.18
fastapi==0.115.4
fastparquet==2024.5.0
flash_attn==2.6.3
flatdict==4.0.1
flint==0.0.1
flytekit==1.13.11
foo==.1
ftfy==6.3.1
fugashi==1.3.2
g2p_en==2.1.0
galore_torch==1.0
gdown==5.2.0
gguf==0.10.0
gluonnlp==0.10.0
gmpy2==2.2.1
gradio==5.4.0
h2==4.1.0
h5py==3.12.1
haiku==0.01
hf_transfer==0.1.8
HTMLParser==0.0.2
hypothesis==6.115.5
importlib_metadata==8.2.0
importlib_resources==6.4.5
InquirerPy==0.3.4
intel_extension_for_pytorch==2.4.0
ipadic==1.0.0
ipython==8.12.3
ipywidgets==8.1.5
isal==1.7.1
jax==0.4.35
jaxlib==0.4.35
jieba==0.42.1
jnius==1.1.0
js==1.0
kenlm==0.2.0
keras_nlp==0.17.0
kerchunk==0.2.6
keyring==25.5.0
laion_clap==1.1.6
librosa==0.10.2.post1
lockfile==0.12.2
lomo_optim==0.1.1
lxml==5.3.0
lz4==4.3.3
lzmaffi==0.3.0
mamba_ssm==2.2.2
Mask2Former==0.1.5
matchpy==0.5.5
megatron==0.5.1
mlflow==2.17.1
mock==5.1.0
monkeytype==23.3.0
moviepy==1.0.3
mtrand==0.1
MultiScaleDeformableAttention==1.0
multiset==3.2.0
mxnet==1.9.1
mypy==1.13.0
natten==0.17.1
neptune==1.12.0
nltk==3.9.1
numarray==1.5.1
Numeric==24.2
omegaconf==2.3.0
onnxscript==0.1.0.dev20241029
openfold==0.0.1
optimum==1.23.2
optree==0.13.0
optuna==4.0.0
ordereddict==1.1
pandas==2.2.3
paramiko==3.5.0
phonemizer==3.3.0
pickle5==0.0.12
Pillow==11.0.0
pretty_midi==0.2.10
protobuf==5.28.3
psutil==6.0.0
py3nvml==0.2.7
pyarrow==16.0.0
pycocotools==2.0.8
pycosat==0.6.6
pyctcdecode==0.5.0
pydantic==2.9.2
pydot==3.0.2
pydy==0.7.1
pygit2==1.16.0
pyglet==2.0.18
pygraphviz==1.14
PyInstaller==6.11.0
pymc==5.17.0
pymc3==3.11.6
pynvml==11.5.3
pyodide==0.0.2
pyOpenSSL==24.2.1
pysat==3.2.1
pytesseract==0.3.13
pytest==8.3.3
pythainlp==5.0.4
python_Levenshtein==0.26.1
pytorch_lightning==2.4.0
pytorch_quantization==2.2.1
pytz==2024.1
railroad==0.5.0
ray==2.38.0
requests_kerberos==0.15.0
rhoknp==1.7.0
rjieba==0.1.11
s3prl==0.4.16
sacremoses==0.1.1
sage==0.0.0
scann==1.3.4
scikit_learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
setuptools_scm==8.1.0
sigopt==8.8.3
simplejson==3.19.3
slack_sdk==3.33.2
smbprotocol==1.14.0
soundfile==0.12.1
spacy==3.8.2
Sphinx==8.1.3
starlette==0.41.2
sudachipy==0.6.8
symengine==0.13.0
t5x==0.0.0
tabulate==0.9.0
tensorboardX==2.6.2.2
tensorflow_hub==0.16.1
tensorflow_probability==0.24.0
tensorflow_text==2.18.0
tensorstore==0.1.67
tf2onnx==1.16.1
tf_keras==2.18.0
theano==1.0.5
thread==2.0.5
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.11
tokenizer==3.4.5
toml==0.10.2
torch_geometric==2.6.1
torch_npu==2.4.0
torch_scatter==2.1.2
torch_xla==2.5.0
torcharrow==0.1.0
torchaudio==2.5.0
torchrec==1.0.0
torchvision==0.20.0
tornado==6.4.1
trajectory==0.1.1
truststore==0.10.0
ujson==5.10.0
unidic==1.1.0
unidic_lite==1.0.8
urllib3_secure_extra==0.1.0
uvicorn==0.32.0
vissl==0.1.6
wandb==0.18.5
wget==3.2
xformers==0.0.28.post2
xmlrpclib==1.0.1
xmlrunner==1.7.7
yarl==1.17.0
z3==0.2.0
zstandard==0.23.0
